% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imteractive.R
\name{imteractiveOutput}
\alias{imteractiveOutput}
\title{Widget output function for use in Shiny}
\usage{
imteractiveOutput(outputId, width = "100\%", height = "500px")
}
\description{
Widget output function for use in Shiny
}
