% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imteractive.R
\name{renderImteractive}
\alias{renderImteractive}
\title{Widget render function for use in Shiny}
\usage{
renderImteractive(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}
