% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imteractive.R
\name{renderimteractive}
\alias{renderimteractive}
\title{Widget render function for use in Shiny}
\usage{
renderimteractive(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}
